Redundant commit, might contain useful reference
This file contains code that I was writing under the assumption that nothing had been started for Program #3 yet. With that program already seeming to be well on it's way, I'd be fine with adding proper comments/debugging it's content instead. I'm used to a more collaborative/decentralized structure for working on assignments like this, so apologies for being unawares until about now, after I had finished the skeleton up.