Very WIP - panels + buttons
I gutted most of Program #4 earlier today and began the process of repurposing pieces of it (and code from #3) to incorporate all 3 layout managers. The UI requires a scrollbar on either side (speed, size) with three buttons in the middle, so keeping the dimensions of the gridbag layout inline with that should be the best bet. A fresh version of OBJC can likely be inserted in for use with <canvas> (should be renamed <screen>), though most of the I/O, screen resizing, and other elements won't change much outside of Ball() itself. Since all of the mouse listeners are meant to be added to Ball(), I have not incorporated any of them as of yet. The thing that I'm assuming will take the most time is mouse-dragging/click sensing, and the ball's actual collision. For functions like resizing the canvas section of the screen, it might be wise to store whatever the smallest/largest values of the rectangles in either direction are, since it saves having to iterate over the entire list to fetch those values every resize. I should have more time tomorrow, so I'll try to finish the control panel and re-incorporate a canvas, so the actual cerebral bit can begin.

If any of you have code that does similar/more, it would be best to replace this with that; if anything to push things further along.