Revised Collision Code
I ended up rewriting the collision code entirely to be a much simpler implementation, that is more robust and overall much easier to follow. It uses masking on the current velocity (if velocity needs to be changed, it is multiplied by -1. This is repeated for every angle the ball can collide with a rectangle, to create a new velocity vector that represents a new, opposite movement direction), and sets Ball's X and Y instead of incrementing them before running any calculations. I ended up gutting a lot of for-loop code and two other methods to do this, but the code appears to have the exact same features overall.