Menus, Hotkeys, and Resizing
I added all of the functionality needed for the menus to work as intended; though the size/speed values could be tweaked to better match the program the professor is running.

I also fixed a bug where the cannon would not move to the proper place if the screen was resized (though, is this program meant to be resized at all?) until the cannon's angle was changed.

I tried to fix a new problem that was causing the main ball object to destroy the red border, instead of bouncing off of it, but I had no luck growing/shrinking the physics objects. The locations of the screen borders probably just need to be offset by one pixel in the relevant directions, instead of the grow/shrink I was trying.

The way I get the menus to talk with objects/the Frame() class is through a string lookup table, which in less elegant terms is a switch-case that takes the name of a menu item, and assigns it to certain return values. This is done instead of an object-lookup table, since it's much easier to manage the CheckboxMenuItems when they all are toggled in bulk instead of individually. Items left to work on include:
- Cannon Ball,
- Gravity Values,
- Score Values,
- Reset Function,
- Dialogue Box.