Brief Time Update
Adds the time equation used to keep (roughly) track of how many seconds have passed since the program began running. The variables of note are time_ms (current millisecond)  and time_s (seconds passed). Whenever a second has passed, time_ms is reset to zero, and is post-incremented to one for the next iteration of the loop.