GUI Skeleton
I gutted Program #6, and reused the code there to the best of my abilities to reproduce the interface shown in the .pdf description the professor provided for the Chat.java program. Theoretically, all that would need to be written from here is the actual listening code (which the professor gives in lessons related to the program). I had some issues with the width/height of elements, since 99% of the elements are added via control_panel's gridbaglayout, with only the chat messages window being added by itself. I managed to make the program look better by just changing the default height to be 500 instead of 400, as a fix (since the weird sizes only show up around 450 and below).

There's a laundry list of listeners that could probably be removed, though leaving them on the GUIChat() class doesn't seem to harm anything for the moment. I removed all of the logic from the listeners, since none of it really could be reused for this program. 

I also already added the logic to accept a string as the server/port from the command line, whenever initially launching the program (though obviously the string isn't validated/used for anything yet).