Nearly Finished Revision
The only thing I haven't implemented is the port parameter, mostly since it isn't actually mentioned on the rubrics for the program. I'm almost positive a lesson asked for the host/port to be able to be assigned from the command line, as an argument, but still. I'll probably just dump the string into portField, since that makes the most sense. I fixed several I/O visual bugs, unified how the error messages are written, and added other cosmetic changes like the program's header always stating the mode it currently is in. I also added functionality to the send button, which had been neglected in previous versions, and removed a lot of redundant chat focus calls.